+++ Classes/ViewHelpers/Format/TrimViewHelper.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace T3v\T3vBase\ViewHelpers\Format;
+
+use FluidTYPO3\Vhs\ViewHelpers\Format\TrimViewHelper as VhsTrimViewHelper;
+
+/**
+ * The trim view helper class.
+ *
+ * @package T3v\T3vBase\ViewHelpers\Format
+ */
+class TrimViewHelper extends VhsTrimViewHelper
+{
+}