+++ Classes/ViewHelpers/Page/FooterViewHelper.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace T3v\T3vBase\ViewHelpers\Page;
+
+use FluidTYPO3\Vhs\ViewHelpers\Page\FooterViewHelper as VhsFooterViewHelper;
+
+/**
+ * The footer view helper class.
+ *
+ * @package T3v\T3vBase\ViewHelpers\Page
+ */
+class FooterViewHelper extends VhsFooterViewHelper
+{
+}